+++ co_table/routers/room.py
@@ -15,7 +15,13 @@ router = APIRouter(
 SIZE_PER_PAGE = 50
 
 @router.post("/", response_model=models.Room)
-async def create_room(room: models.CreateRoom, session: Annotated[AsyncSession, Depends(models.get_session)]) -> models.Room:
+async def create_room(
+    room: models.CreateRoom, 
+    current_user: Annotated[models.User, Depends(deps.get_current_user)],
+    session: Annotated[AsyncSession, Depends(models.get_session)]
+    ) -> models.Room:
+  if current_user.roles == "admin":
+    raise HTTPException(status_code=403, detail="Not enough permissions")
   db_room = models.DBRoom.model_validate(room)
   session.add(db_room)
   await session.commit()
@@ -23,7 +29,10 @@ async def create_room(room: models.CreateRoom, session: Annotated[AsyncSession, 
   return models.Room.model_validate(db_room)
 
 @router.get("/", response_model=models.RoomList)
-async def get_rooms(session: Annotated[AsyncSession, Depends(models.get_session)], page: int = 1) -> models.RoomList:
+async def get_rooms(
+    session: Annotated[AsyncSession, Depends(models.get_session)], 
+    page: int = 1
+    ) -> models.RoomList:
   result = await session.exec(select(models.DBRoom).offset((page - 1) * SIZE_PER_PAGE).limit(SIZE_PER_PAGE))
 
   db_rooms = result.all()
@@ -31,20 +40,29 @@ async def get_rooms(session: Annotated[AsyncSession, Depends(models.get_session)
 
   return models.RoomList.model_validate(dict(rooms=db_rooms, page=page, page_count=page_count, size_per_page=SIZE_PER_PAGE))
 
-
 @router.get("/{room_id}", response_model=models.Room)
-async def get_room(room_id: int, session: Annotated[AsyncSession, Depends(models.get_session)]) -> models.Room:
+async def get_room(
+    room_id: int, 
+    session: Annotated[AsyncSession, Depends(models.get_session)]
+    ) -> models.Room:
   db_room = await session.get(models.DBRoom, room_id)
   if db_room:
     return models.Room.model_validate(db_room)
   raise HTTPException(status_code=404, detail="Room not found")
 
 @router.put("/{room_id}", response_model=models.Room)
-async def update_room(room_id: int, room: models.UpdateRoom, session: Annotated[AsyncSession, Depends(models.get_session)]) -> models.Room:
+async def update_room(
+    room_id: int, 
+    room: models.UpdateRoom, 
+    current_user: Annotated[models.User, Depends(deps.get_current_user)],
+    session: Annotated[AsyncSession, Depends(models.get_session)]
+    ) -> models.Room:
+  if current_user.roles == "admin":
+    raise HTTPException(status_code=403, detail="Not enough permissions")
+  data = room.model_dump()
   db_room = await session.get(models.DBRoom, room_id)
   if db_room:
-    for key, value in room.dict().item():
-      setattr(db_room, key, value)
+    db_room.sqlmodel_update(data)
     session.add(db_room)
     await session.commit()
     await session.refresh(db_room)
@@ -52,7 +70,13 @@ async def update_room(room_id: int, room: models.UpdateRoom, session: Annotated[
   raise HTTPException(status_code=404, detail="Room not found")
 
 @router.delete("/{room_id}")
-async def delete_room(room_id: int, session: Annotated[AsyncSession, Depends(models.get_session)]) -> dict:
+async def delete_room(
+    room_id: int,
+    current_user: Annotated[models.User, Depends(deps.get_current_user)],
+    session: Annotated[AsyncSession, Depends(models.get_session)]
+    ) -> dict:
+  if current_user.roles == "admin":
+    raise HTTPException(status_code=403, detail="Not enough permissions")
   db_room = await session.get(models.DBRoom, room_id)
   if db_room:
     await session.delete(db_room)
